move all includes into UserCompactTransformer
