+++ app/Transformers/UserTransformer.php
@@ -22,18 +22,6 @@ class UserTransformer extends UserCompactTransformer
         'is_restricted',
     ];
 
-    protected $permissions = [
-        'friends' => 'IsNotOAuth',
-        'is_admin' => 'IsNotOAuth',
-        'is_bng' => 'IsNotOAuth',
-        'is_full_bn' => 'IsNotOAuth',
-        'is_gmt' => 'IsNotOAuth',
-        'is_limited_bn' => 'IsNotOAuth',
-        'is_moderator' => 'IsNotOAuth',
-        'is_nat' => 'IsNotOAuth',
-        'is_restricted' => 'IsNotOAuth',
-    ];
-
     public function transform(User $user)
     {
         $result = parent::transform($user);
@@ -68,45 +56,5 @@ class UserTransformer extends UserCompactTransformer
             'twitter' => $user->user_twitter,
             'website' => $user->user_website,
         ]);
-    }
-
-    public function includeIsAdmin(User $user)
-    {
-        return $this->primitive($user->isAdmin());
-    }
-
-    public function includeIsBng(User $user)
-    {
-        return $this->primitive($user->isBNG());
-    }
-
-    public function includeIsFullBn(User $user)
-    {
-        return $this->primitive($user->isFullBN());
-    }
-
-    public function includeIsGmt(User $user)
-    {
-        return $this->primitive($user->isGMT());
-    }
-
-    public function includeIsLimitedBn(User $user)
-    {
-        return $this->primitive($user->isLimitedBN());
-    }
-
-    public function includeIsModerator(User $user)
-    {
-        return $this->primitive($user->isModerator());
-    }
-
-    public function includeIsNat(User $user)
-    {
-        return $this->primitive($user->isNAT());
-    }
-
-    public function includeIsRestricted(User $user)
-    {
-        return $this->primitive($user->isRestricted());
     }
 }
